las solicitudes ya apreceren bien, pero no funciona el aceptar y rechazar y tampoco se borran cuando se apretan esos botones. ademas falta el agregar user nuevo. Cuando se aprete aceptar tiene que agregarse al proyecto como si es admin o no y ademas agregarlo y eliminarse de las solicitudes pendientes, si es rechazar solo eliminar de solicitudes pendientes y refrescar

+++ WebFinalProjectNicolasScandolo3/Models/RequestToProyecto.cs
@@ -17,7 +17,7 @@ namespace WebFinalProjectNicolasScandolo3.Models
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
         public int IdProyecto { get; set; }
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-        public int IdUsuario { get; set; }
+        public long IdUsuario { get; set; }
 
     }
 }
